Refactor texteditor into its own module.
